Fixes openalea-incubator/csMTG#17 openalea-incubator/csMTG#18 openalea-incubator/csMTG#19
- Deleted function NbVertices() and kept only Count().
- Change the way a new id is attributed.
- Remove the algorithm for generating a random tree into a new class
(Algorithm).
- Create separate tests for the new class Algorithm.
