+++ code/Tree.cs
@@ -110,27 +110,21 @@ namespace csMTG
             int count = parent.Count();
             return count;
         }
-
-        /// <summary>
-        /// Counts the number of vertices in the tree.
-        /// </summary>
-        /// <returns> Returns the total number of vertices. </returns>
-        public int NbVertices()
-        {
-            int nbVertices = parent.Count();
-            return nbVertices;
-        }
+        
 
         /// <summary>
         /// Function which generates a unique Id.
         /// </summary>
         private int NewId()
         {
+            id = parent.Keys.Max() + 1;
+            /*
             do
             {
                 id++;
             }
             while (parent.ContainsKey(id));
+            */
             return id;
         }
 
@@ -250,49 +244,6 @@ namespace csMTG
             }
         }
 
-        /// <summary>
-        ///  Generates a random tree with a specified number of vertices and children per vertex
-        /// </summary>
-        /// <param name="nbVertices"> The number of tree's vertices </param>
-        /// <param name="nbChildren"> The maximum number for a vertex </param>
-        /// <returns> Returns the generated tree </returns>
-        public Tree RandomTree(Tree t, int nbVertices, int nbChildren = 4)
-        {
-            int childrenToAdd;
-
-            List<int> randomStack = new List<int>();
-            Random r = new Random();
-            Random r2 = new Random();
-            randomStack.Add(t.root);
-            nbVertices--;
-
-            while (nbVertices > 0)
-            {
-                // Set a random number of children to add within the specified range.
-                
-                int randomInt = r.Next(1, nbChildren);
-                childrenToAdd = Math.Min(randomInt, nbVertices);
-
-                // Choose a random vertex among those in the stack of potential parents
-                
-                int randomNumber = r2.Next(0, randomStack.Count()-1);
-                int randomVertex = randomStack[randomNumber];
-
-                // Add the specified number of children to the random parent
-                for (int i = 0; i < childrenToAdd; ++i)
-                {
-                    int newChild = AddChild(randomVertex);
-                    randomStack.Add(newChild);
-                    nbVertices--;
-                }
-
-                randomStack.Remove(randomVertex);
-
-            }
-
-            return t;
-        }
-        
 
         static void Main(String[] args)   
         {
