+++ codeTests/AlgorithmTests.cs
@@ -1,0 +1,16 @@
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using csMTG;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using csMTG.Algorithm;
+
+namespace csMTG.Tests
+{
+    [TestClass()]
+    public class AlgorithmTests
+    {
+        }
+}
