+++ test/TreeTests.cs
@@ -227,64 +227,6 @@ namespace csMTG.Tests
         }
         #endregion
 
-        #region Tests of the random tree generator
-        [TestMethod()]
-        public void RandomTree_NumberOfVerticesCreated_SameAsParameter()
-        {
-            Tree t = new Tree();
-
-            int numberOfExpectedVertices = 50;
-
-            t = t.RandomTree(t, numberOfExpectedVertices);
-
-            Assert.AreEqual(numberOfExpectedVertices, t.NbVertices());
-            
-        }
-
-        [TestMethod()]
-        public void RandomTree_NumberOfChildrenRespected_LessOrEqualToTheParameter()
-        {
-            Tree t = new Tree();
-
-            int maximumNbChildren = 5;
-
-            t = t.RandomTree(t, 50, maximumNbChildren);
-
-            foreach(int keyId in t.children.Keys)
-            {
-                Assert.IsTrue(t.NbChildren(keyId) <= maximumNbChildren);
-            }
-
-        }
-
-        [TestMethod()]
-        public void RandomTree_DoChildrenAndParentCorrespond_CoherenceBetweenParentAndChildren()
-        {
-            Tree t = new Tree();
-            
-            t = t.RandomTree(t, 50, 5);
-            
-            foreach(int childId in t.parent.Keys)
-            {
-                if(childId != 0)
-                {
-                    int? parentId = t.Parent(childId);
-                    Assert.IsNotNull(parentId);
-
-                    CollectionAssert.Contains(t.Children((int)parentId), childId);
-                }
-            }
-
-            foreach(int parentId in t.children.Keys)
-            {
-                foreach(int childId in t.Children(parentId))
-                {
-                    Assert.AreEqual(t.Parent(childId), parentId);
-                }
-            }
-
-        }
-
-        #endregion
+        
     }
 }
